optimize with f32

+++ crates/classy/src/core/activation/mod.rs
@@ -8,7 +8,7 @@ pub enum Activation {
 }
 
 impl Activation {
-    pub fn call_on_all(&self, h: Array2<f64>) -> Array2<f64> {
+    pub fn call_on_all(&self, h: Array2<f32>) -> Array2<f32> {
         match self {
             Self::Linear => h,
             Self::Sigmoid => h.map(|x| sigmoid(*x)),
@@ -17,7 +17,7 @@ impl Activation {
                 let mut res = Array2::zeros((h.nrows(), h.ncols()));
                 for (mut res_row, h_row) in res.axis_iter_mut(Axis(0)).zip(h.axis_iter(Axis(0))) {
                     let exp_values = h_row.map(|v| v.exp());
-                    let sum_exp: f64 = exp_values.iter().sum();
+                    let sum_exp: f32 = exp_values.iter().sum();
                     res_row.assign(&(exp_values / sum_exp));
                 }
                 res
@@ -26,6 +26,6 @@ impl Activation {
     }
 }
 
-fn sigmoid(x: f64) -> f64 {
+fn sigmoid(x: f32) -> f32 {
     1.0 / (1.0 + (-x).exp())
 }
