+++ crates/classy/src/core/loss/crossentropy.rs
@@ -1,9 +1,9 @@
 use ndarray::{Array1, Array2, Axis};
 
-const EPSILON: f64 = 1e-15;
+const EPSILON: f32 = 1e-15;
 
-pub fn bin_cross_entropy(y: &Array2<f64>, y1: &Array2<f64>) -> Array2<f64> {
-    let res: Array1<f64> = y1
+pub fn bin_cross_entropy(y: &Array2<f32>, y1: &Array2<f32>) -> Array2<f32> {
+    let res: Array1<f32> = y1
         .iter()
         .zip(y.iter())
         .map(|(y1_i, y_i)| {
@@ -18,18 +18,18 @@ pub fn bin_cross_entropy(y: &Array2<f64>, y1: &Array2<f64>) -> Array2<f64> {
         .to_owned()
 }
 
-pub fn bin_cross_entropy_d(y: &Array2<f64>, y1: &Array2<f64>) -> Array2<f64> {
+pub fn bin_cross_entropy_d(y: &Array2<f32>, y1: &Array2<f32>) -> Array2<f32> {
     y1 - y
 }
 
-pub fn cross_entropy(y: &Array2<f64>, y1: &Array2<f64>) -> Array2<f64> {
-    (-y * (&y1.map(|x| x.max(EPSILON).min(1f64 - EPSILON).ln())))
+pub fn cross_entropy(y: &Array2<f32>, y1: &Array2<f32>) -> Array2<f32> {
+    (-y * (&y1.map(|x| x.max(EPSILON).min(1f32 - EPSILON).ln())))
         .sum_axis(Axis(1))
         .to_shape((y.nrows(), 1))
         .unwrap()
         .to_owned()
 }
 
-pub fn cross_entropy_d(y: &Array2<f64>, y1: &Array2<f64>) -> Array2<f64> {
+pub fn cross_entropy_d(y: &Array2<f32>, y1: &Array2<f32>) -> Array2<f32> {
     y1 - y
 }
