+++ crates/classy/src/core/solver/ols.rs
@@ -7,9 +7,9 @@ use ndarray::{Array2, Axis};
 pub struct OrdinaryLeastSquares;
 
 impl OrdinaryLeastSquares {
-    pub fn train(&self, data: &Array2<f64>, targets: &Array2<f64>, silent: bool) -> Array2<f64> {
+    pub fn train(&self, data: &Array2<f32>, targets: &Array2<f32>, silent: bool) -> Array2<f32> {
         let y_mean = targets.mean_axis(Axis(0));
-        let mut sst = 0f64;
+        let mut sst = 0f32;
         for row in targets.axis_iter(Axis(0)) {
             sst += (row - &y_mean).map(|x| x.powi(2)).sum();
         }
@@ -24,20 +24,20 @@ impl OrdinaryLeastSquares {
             None => panic!("Unable to invert XtX"),
         };
         // OLS Weights = (XtX)' * XtY
-        let weights: Array2<f64> = xtx_inverse * xty;
+        let weights: Array2<f32> = xtx_inverse * xty;
 
         // Calculate R2 only if `silent` is set to false
         if !silent {
             let sse = (targets - (data * &weights))
                 .map(|x| x.powi(2))
                 .sum();
-            let r2 = 1f64 - (sse / sst);
+            let r2 = 1f32 - (sse / sst);
             println!("R2 Score: {}", r2);
         }
 
         weights
     }
-    pub fn predict(&self, data: &Array2<f64>, weights: &Array2<f64>) -> Array2<f64> {
+    pub fn predict(&self, data: &Array2<f32>, weights: &Array2<f32>) -> Array2<f32> {
         data * weights
     }
 }
